requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
requests
jupyter
pyyaml
pytest
